refactor: rename HeimOperations to PoolOperations
