+++ src/components/PoolOperations/SelectOperatorMobile/styles.ts
